% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_functions.R
\name{delete}
\alias{delete}
\title{Removes a survey from view in all user GET requests.}
\usage{
delete(surveyUrlId)
}
\arguments{
\item{surveyUrlId}{External URL ID for the survey}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/surveys
\item https://www.googleapis.com/auth/userinfo.email
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/surveys, https://www.googleapis.com/auth/userinfo.email)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{}{Google Documentation}
}
