% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_objects.R
\name{ResultsMask}
\alias{ResultsMask}
\title{ResultsMask Object}
\usage{
ResultsMask(fields = NULL, projection = NULL)
}
\arguments{
\item{fields}{No description}

\item{projection}{No description}
}
\value{
ResultsMask object
}
\description{
ResultsMask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
