% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_functions.R
\name{start}
\alias{start}
\title{Begins running a survey.}
\usage{
start(SurveysStartRequest, resourceId)
}
\arguments{
\item{SurveysStartRequest}{The \link{SurveysStartRequest} object to pass to this method}

\item{resourceId}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/surveys
\item https://www.googleapis.com/auth/userinfo.email
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/surveys, https://www.googleapis.com/auth/userinfo.email)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{}{Google Documentation}

Other SurveysStartRequest functions: \code{\link{SurveysStartRequest}}
}
