% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_objects.R
\name{Survey}
\alias{Survey}
\title{Survey Object}
\usage{
Survey(audience = NULL, cost = NULL, customerData = NULL,
  description = NULL, owners = NULL, questions = NULL,
  rejectionReason = NULL, state = NULL, surveyUrlId = NULL,
  title = NULL, wantedResponseCount = NULL)
}
\arguments{
\item{audience}{Targeting-criteria message containing demographic information}

\item{cost}{Cost to run the survey and collect the necessary number of responses}

\item{customerData}{Additional information to store on behalf of the API consumer and associate with this question}

\item{description}{Text description of the survey}

\item{owners}{List of email addresses for survey owners}

\item{questions}{List of questions defining the survey}

\item{rejectionReason}{Reason for the survey being rejected}

\item{state}{State that the survey is in}

\item{surveyUrlId}{Unique survey ID, that is viewable in the URL of the Survey Creator UI}

\item{title}{Optional name that will be given to the survey}

\item{wantedResponseCount}{Number of responses desired for the survey}
}
\value{
Survey object
}
\description{
Survey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Representation of an individual survey object.
}
\seealso{
Other Survey functions: \code{\link{insert}},
  \code{\link{update}}
}
