% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_objects.R
\name{SurveyAudience}
\alias{SurveyAudience}
\title{SurveyAudience Object}
\usage{
SurveyAudience(ages = NULL, country = NULL, countrySubdivision = NULL,
  gender = NULL, languages = NULL, mobileAppPanelId = NULL,
  populationSource = NULL)
}
\arguments{
\item{ages}{Optional list of age buckets to target}

\item{country}{Required country code that surveys should be targeted to}

\item{countrySubdivision}{Country subdivision (states/provinces/etc) that surveys should be targeted to}

\item{gender}{Optional gender to target}

\item{languages}{Language code that surveys should be targeted to}

\item{mobileAppPanelId}{Key for predefined panel that causes survey to be sent to a predefined set of Opinion Rewards App users}

\item{populationSource}{Online population source where the respondents are sampled from}
}
\value{
SurveyAudience object
}
\description{
SurveyAudience Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifications for the target audience of a survey run through the API.
}
