% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_objects.R
\name{SurveyRejection}
\alias{SurveyRejection}
\title{SurveyRejection Object}
\usage{
SurveyRejection(explanation = NULL, type = NULL)
}
\arguments{
\item{explanation}{A human-readable explanation of what was wrong with the survey}

\item{type}{Which category of rejection this was}
}
\value{
SurveyRejection object
}
\description{
SurveyRejection Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message representing why the survey was rejected from review, if it was.
}
