% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_objects.R
\name{SurveyResults}
\alias{SurveyResults}
\title{SurveyResults Object}
\usage{
SurveyResults(status = NULL, surveyUrlId = NULL)
}
\arguments{
\item{status}{Human readable string describing the status of the request}

\item{surveyUrlId}{External survey ID as viewable by survey owners in the editor view}
}
\value{
SurveyResults object
}
\description{
SurveyResults Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Reference to the current results for a given survey.
}
