% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_functions.R
\docType{package}
\name{surveys_googleAuthR}
\alias{surveys_googleAuthR}
\alias{surveys_googleAuthR-package}
\title{Surveys API
Creates and conducts surveys, lists the surveys that an authenticated user owns, and retrieves survey results and information about specified surveys.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:18:13
filename: /Users/mark/dev/R/autoGoogleAPI/googlesurveysv2.auto/R/surveys_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/surveys
\item https://www.googleapis.com/auth/surveys.readonly
\item https://www.googleapis.com/auth/userinfo.email
}
}
