% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_objects.R
\name{SurveysStartRequest}
\alias{SurveysStartRequest}
\title{SurveysStartRequest Object}
\usage{
SurveysStartRequest(maxCostPerResponseNanos = NULL)
}
\arguments{
\item{maxCostPerResponseNanos}{Threshold to start a survey automically if the quoted prices is less than or equal to this value}
}
\value{
SurveysStartRequest object
}
\description{
SurveysStartRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SurveysStartRequest functions: \code{\link{start}}
}
