% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_objects.R
\name{TokenPagination}
\alias{TokenPagination}
\title{TokenPagination Object}
\usage{
TokenPagination(nextPageToken = NULL, previousPageToken = NULL)
}
\arguments{
\item{nextPageToken}{No description}

\item{previousPageToken}{No description}
}
\value{
TokenPagination object
}
\description{
TokenPagination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
